pyproject.toml
[project]
name = "paperbot"
version = "1.0.0"
description = "Journal RSS → Crossref → SQLite → Zotero pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "PaperBot Contributors" }
]
keywords = ["rss", "zotero", "crossref", "papers", "academic", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "feedparser>=6.0.0",
    "requests>=2.28.0",
    "python-dateutil>=2.8.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "pyzotero>=1.5.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
paperbot = "paperbot.cli:main"

[project.urls]
Homepage = "https://github.com/wonjuncio/paperbot"
Repository = "https://github.com/wonjuncio/paperbot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["paperbot"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=paperbot"
